Renamed file to reset_estimator.py
