+++ examples/autonomy/autonomous_sequence_high_level_compressed.py
@@ -39,7 +39,7 @@ from cflib.crazyflie.mem import CompressedStart
 from cflib.crazyflie.mem import MemoryElement
 from cflib.crazyflie.syncCrazyflie import SyncCrazyflie
 from cflib.utils import uri_helper
-from cflib.utils.ResetEstimator import reset_estimator
+from cflib.utils.reset_estimator import reset_estimator
 
 # URI to the Crazyflie to connect to
 uri = uri_helper.uri_from_env(default='radio://0/80/2M/E7E7E7E7E7')
