+++ examples/lighthouse/lighthouse_openvr_grab.py
@@ -11,7 +11,7 @@ import cflib.crtp
 from cflib.crazyflie import Crazyflie
 from cflib.crazyflie.log import LogConfig
 from cflib.crazyflie.syncCrazyflie import SyncCrazyflie
-from cflib.utils.ResetEstimator import reset_estimator
+from cflib.utils.reset_estimator import reset_estimator
 
 # URI to the Crazyflie to connect to
 uri = 'radio://0/80/2M/E7E7E7E7E7'
