+++ examples/positioning/flowsequenceSync.py
@@ -35,7 +35,7 @@ import cflib.crtp
 from cflib.crazyflie import Crazyflie
 from cflib.crazyflie.syncCrazyflie import SyncCrazyflie
 from cflib.utils import uri_helper
-from cflib.utils.ResetEstimator import reset_estimator
+from cflib.utils.reset_estimator import reset_estimator
 
 URI = uri_helper.uri_from_env(default='radio://0/80/2M/E7E7E7E7E7')
